add node usage
