+++ JSON/node-color.json
@@ -1,0 +1,10 @@
+{
+    "0":"30",
+    "1":"31",
+    "2":"32",
+    "3":"33",
+    "4":"34",
+    "5":"35",
+    "6":"36",
+    "7":"37"
+}
