+++ JSON/node-font-property.json
@@ -1,0 +1,8 @@
+{
+    "0":"99",
+    "1":"01",
+    "4":"04",
+    "5":"05",
+    "7":"07",
+    "8":"08"
+}
