+++ JSON/nodeBackground-color.json
@@ -1,0 +1,10 @@
+{
+    "0":"40",
+    "1":"41",
+    "2":"42",
+    "3":"43",
+    "4":"44",
+    "5":"45",
+    "6":"46",
+    "7":"47"
+}
